Added possibility to set GitHub or BitBucket account for cloning submodules

+++ steps/serverless/build-bash.sh
@@ -15,12 +15,6 @@ echo ""
 if [ -z "$BLUEMIX_ENV" ]; then
   echo "Missing environment variable BLUEMIX_ENV"
   exit 1
-elif [ -< "$BITBUCKET_USERNAME" ]; then
-  echo "Missing environment variable BITBUCKET_USERNAME"
-  exit 1
-elif [ -< "$GITHUB_USERNAME" ]; then
-  echo "Missing environment variable GITHUB_USERNAME"
-  exit 1
 elif [ -z "$GIT_URL" ]; then
   echo "Missing environment variable GIT_URL"
   exit 1
@@ -52,13 +46,21 @@ cat $WORKSPACE/package.json
 echo ""
 echo "Clone Submodules:"
 
-cat .gitmodules | sed -r s/bitbucket.org/$BITBUCKET_USERNAME@bitbucket.org/ > .gitmodules.tmp
-rm .gitmodules
-mv .gitmodules.tmp .gitmodules
+if [ -z "$BITBUCKET_USERNAME" ]; then
+  if [ -z "$BITBUCKET_PASSWORD" ]; then
+    cat .gitmodules | sed -r s/bitbucket.org/$BITBUCKET_USERNAME:$BITBUCKET_PASSWORD@bitbucket.org/ > .gitmodules.tmp
+    rm .gitmodules
+    mv .gitmodules.tmp .gitmodules
+  fi
+fi
 
-cat .gitmodules | sed -r s/github.com/$GITHUB_USERNAME@github.com/ > .gitmodules.tmp
-rm .gitmodules
-mv .gitmodules.tmp .gitmodules
+if [ -z "$GITHUB_USERNAME" ]; then
+  if [ -z "$GITHUB_PASSWORD" ]; then
+    cat .gitmodules | sed -r s/github.com/$GITHUB_USERNAME:$GITHUB_PASSWORD@github.com/ > .gitmodules.tmp
+    rm .gitmodules
+    mv .gitmodules.tmp .gitmodules
+  fi
+fi
 
 git submodule update --init --recursive
